Adding Updates:
Added a Page Controller which makes managing the viewable page easier.
Added Settings Page
Added a way to modify the transparency within the application.
Added Editior Menu in the application menu.
Changed font button location to editior menu.
Changed the look of the TextBox to make it look slightly better when in use.
Fixed the bug of everything not scaling correctly.